Cast to keep the compiler happy

+++ UKFNSubscribeFileWatcher.m
@@ -117,7 +117,7 @@ void    UKFileSubscriptionProc(FNMessage message, OptionBits flags, void *refcon
 									kNilOptions, &subscription );
 		if( err != noErr )
 		{
-			NSLog( @"UKFNSubscribeFileWatcher addPath: %@ failed due to error ID=%d.", path, err );
+			NSLog( @"UKFNSubscribeFileWatcher addPath: %@ failed due to error ID=%d.", path, (int)err );
 			return;
 		}
 		
@@ -240,5 +240,5 @@ void    UKFileSubscriptionProc( FNMessage message, OptionBits flags, void *refco
     if( message == kFNDirectoryModifiedMessage )    // No others exist as of 10.4
         [obj sendDelegateMessage: message forSubscription: subscription];
     else
-        NSLog( @"UKFileSubscriptionProc: Unknown message %d", message );
+        NSLog( @"UKFileSubscriptionProc: Unknown message %d", (unsigned int)message );
 }
